🧹: rename and gather as codegen
