+++ codegen/codegen/directive_comment.go
@@ -1,4 +1,4 @@
-package astmeta
+package codegen
 
 import (
 	"fmt"
