+++ codegen/codegen/directive_comment_test.go
@@ -1,4 +1,4 @@
-package astmeta
+package codegen
 
 import (
 	"go/parser"
