+++ codegen/codegen/parser.go
@@ -1,4 +1,4 @@
-package astmeta
+package codegen
 
 import (
 	"bytes"
