+++ codegen/generator/undgen/gen_patcher.go
@@ -19,7 +19,7 @@ import (
 	"github.com/dave/dst"
 	"github.com/dave/dst/decorator"
 	"github.com/dave/dst/dstutil"
-	"github.com/ngicks/go-codegen/codegen/astmeta"
+	"github.com/ngicks/go-codegen/codegen/codegen"
 	"github.com/ngicks/go-codegen/codegen/imports"
 	"github.com/ngicks/go-codegen/codegen/structtag"
 	"github.com/ngicks/go-codegen/codegen/suffixwriter"
@@ -111,7 +111,7 @@ func GeneratePatcher(
 			// But we are not printing gen decl itself since
 			// it could have multiple specs inside it (type (spec1; spec2;...))
 			// surely at least a spec of them is converted but we can't tell all of them were.
-			buf.WriteString("//" + astmeta.DirectivePrefix + astmeta.DirectiveCommentGenerated + "\n")
+			buf.WriteString("//" + codegen.DirectivePrefix + codegen.DirectiveCommentGenerated + "\n")
 			buf.WriteString(token.TYPE.String())
 			buf.WriteByte(' ')
 			err = printer.Fprint(buf, res.Fset, ts)
@@ -357,7 +357,7 @@ func generateFromValue(
 	printf(
 		`//%s%s
 `,
-		astmeta.DirectivePrefix, astmeta.DirectiveCommentGenerated,
+		codegen.DirectivePrefix, codegen.DirectiveCommentGenerated,
 	)
 	printf(
 		`func (p *%s) FromValue(v %s) {
@@ -453,7 +453,7 @@ func generateToValue(
 	printf(
 		`//%s%s
 `,
-		astmeta.DirectivePrefix, astmeta.DirectiveCommentGenerated,
+		codegen.DirectivePrefix, codegen.DirectiveCommentGenerated,
 	)
 	printf(
 		`func (p %s) ToValue() %s {
@@ -542,7 +542,7 @@ func generateMerge(
 
 	printf(`//%s%s
 `,
-		astmeta.DirectivePrefix, astmeta.DirectiveCommentGenerated,
+		codegen.DirectivePrefix, codegen.DirectiveCommentGenerated,
 	)
 	printf(`func (p %[1]s) Merge(r %[1]s) %[1]s {
 `,
@@ -629,7 +629,7 @@ func generateApplyPatch(
 	printf(
 		`//%s%s
 `,
-		astmeta.DirectivePrefix, astmeta.DirectiveCommentGenerated,
+		codegen.DirectivePrefix, codegen.DirectiveCommentGenerated,
 	) // note this is generated method.
 	printf(
 		`func (p %[1]s) ApplyPatch(v %[2]s) %[2]s {
